+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Hydrogen // Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add a warning about the public instance ending.
+- Add Portuguese Brazil translation to the public instance warning.
+
 ## [0.0.1-alpha.7] - 2024-04-15
 
 ### Added